# public/doc/admin/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Crispy API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Crispy application",
    "contact": {
      "name": "API Support",
      "email": "support@crispy.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4200/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "error": {
            "type": "string",
            "description": "Error details"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/admin/access-token": {
      "get": {
        "tags": [
          "AccessTokens"
        ],
        "summary": "获取Access Token列表",
        "description": "获取Access Token列表，支持分页和搜索",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每页数量"
          },
          {
            "in": "query",
            "name": "app_name",
            "schema": {
              "type": "string"
            },
            "description": "应用名称（模糊搜索）"
          },
          {
            "in": "query",
            "name": "channel",
            "schema": {
              "type": "string"
            },
            "description": "渠道名称（模糊搜索）"
          },
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "description": "Token值（模糊搜索）"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer"
            },
            "description": "状态"
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer"
            },
            "description": "用户ID"
          },
          {
            "in": "query",
            "name": "is_delete",
            "schema": {
              "type": "integer"
            },
            "description": "删除状态"
          },
          {
            "in": "query",
            "name": "update_time",
            "schema": {
              "type": "integer"
            },
            "description": "更新时间戳"
          },
          {
            "in": "query",
            "name": "create_time",
            "schema": {
              "type": "integer"
            },
            "description": "创建时间戳"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AccessTokens"
        ],
        "summary": "创建Access Token",
        "description": "创建新的Access Token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "app_name",
                  "channel"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Token名称"
                  },
                  "app_name": {
                    "type": "string",
                    "description": "应用名称"
                  },
                  "channel": {
                    "type": "string",
                    "description": "渠道名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "Token描述"
                  },
                  "expire_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "过期时间"
                  },
                  "status": {
                    "type": "integer",
                    "description": "状态"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/access-token/{id}": {
      "get": {
        "tags": [
          "AccessTokens"
        ],
        "summary": "获取Access Token详情",
        "description": "根据ID获取Access Token详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Access Token ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Access Token不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AccessTokens"
        ],
        "summary": "更新Access Token",
        "description": "根据ID更新Access Token信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Access Token ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Token名称"
                  },
                  "app_name": {
                    "type": "string",
                    "description": "应用名称"
                  },
                  "channel": {
                    "type": "string",
                    "description": "渠道名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "Token描述"
                  },
                  "expire_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "过期时间"
                  },
                  "status": {
                    "type": "integer",
                    "description": "状态"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AccessTokens"
        ],
        "summary": "删除Access Token",
        "description": "根据ID删除Access Token（软删除）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Access Token ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/ad-items": {
      "get": {
        "tags": [
          "AdItems"
        ],
        "summary": "获取广告项目列表",
        "description": "获取广告项目列表，支持分页和搜索",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每页数量"
          },
          {
            "in": "query",
            "name": "ad_id",
            "schema": {
              "type": "integer"
            },
            "description": "广告ID过滤"
          },
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "广告项目标题（模糊搜索）"
          },
          {
            "in": "query",
            "name": "content",
            "schema": {
              "type": "string"
            },
            "description": "广告项目内容（模糊搜索）"
          },
          {
            "in": "query",
            "name": "image_url",
            "schema": {
              "type": "string"
            },
            "description": "图片URL（模糊搜索）"
          },
          {
            "in": "query",
            "name": "url",
            "schema": {
              "type": "string"
            },
            "description": "链接地址（模糊搜索）"
          },
          {
            "in": "query",
            "name": "method",
            "schema": {
              "type": "string"
            },
            "description": "打开方式"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "integer"
            },
            "description": "排序值"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer"
            },
            "description": "状态过滤"
          },
          {
            "in": "query",
            "name": "is_delete",
            "schema": {
              "type": "integer"
            },
            "description": "删除状态"
          },
          {
            "in": "query",
            "name": "update_time",
            "schema": {
              "type": "integer"
            },
            "description": "更新时间戳"
          },
          {
            "in": "query",
            "name": "create_time",
            "schema": {
              "type": "integer"
            },
            "description": "创建时间戳"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdItems"
        ],
        "summary": "创建广告项目",
        "description": "创建新广告项目",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "ad_id"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "广告项目标题"
                  },
                  "ad_id": {
                    "type": "integer",
                    "description": "所属广告ID"
                  },
                  "content": {
                    "type": "string",
                    "description": "广告项目内容"
                  },
                  "url": {
                    "type": "string",
                    "description": "链接地址"
                  },
                  "image": {
                    "type": "string",
                    "description": "图片地址"
                  },
                  "status": {
                    "type": "integer",
                    "default": 1,
                    "description": "状态 (-10-禁用, 10-启用)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/ad-items/{id}": {
      "get": {
        "tags": [
          "AdItems"
        ],
        "summary": "获取广告项目详情",
        "description": "根据ID获取广告项目详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "广告项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "广告项目不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdItems"
        ],
        "summary": "更新广告项目",
        "description": "根据ID更新广告项目信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "广告项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "广告项目标题"
                  },
                  "ad_id": {
                    "type": "integer",
                    "description": "所属广告ID"
                  },
                  "content": {
                    "type": "string",
                    "description": "广告项目内容"
                  },
                  "url": {
                    "type": "string",
                    "description": "链接地址"
                  },
                  "image": {
                    "type": "string",
                    "description": "图片地址"
                  },
                  "status": {
                    "type": "integer",
                    "description": "状态 (-10-禁用, 10-启用)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "广告项目不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdItems"
        ],
        "summary": "删除广告项目",
        "description": "根据ID删除广告项目（软删除）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "广告项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "广告项目不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/page-cache/stats": {
      "get": {
        "tags": [
          "PageCache"
        ],
        "summary": "获取页面缓存统计信息",
        "description": "获取页面缓存的统计数据",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/page-cache/memory/list": {
      "get": {
        "tags": [
          "PageCache"
        ],
        "summary": "获取内存缓存列表",
        "description": "获取所有页面内存缓存的列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/page-cache/memory/cleanup": {
      "post": {
        "tags": [
          "PageCache"
        ],
        "summary": "清理过期内存缓存",
        "description": "清理所有过期的页面内存缓存",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "清理成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/page-cache/database/list": {
      "get": {
        "tags": [
          "PageCache"
        ],
        "summary": "获取数据库缓存列表",
        "description": "获取所有页面数据库缓存的列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/page-cache/database/cleanup": {
      "post": {
        "tags": [
          "PageCache"
        ],
        "summary": "清理过期数据库缓存",
        "description": "清理所有过期的页面数据库缓存",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "清理成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}